+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add option to toggle panel placement between bottom and right
 - Add `:margin-element` option to automatically set margin-bottom/right on some
 element, so that panel doesn't go over the application content
+- Store open paths on state tree to local storage
 
 # 0.2.1 (2020-03-06)